+++ magik-checks/src/main/java/nl/ramsolutions/sw/magik/checks/checks/WarnedCallCheck.java
@@ -4,6 +4,7 @@ import com.sonar.sslr.api.AstNode;
 import java.util.Arrays;
 import java.util.Set;
 import java.util.stream.Collectors;
+import nl.ramsolutions.sw.magik.analysis.helpers.MethodInvocationNodeHelper;
 import nl.ramsolutions.sw.magik.api.MagikGrammar;
 import nl.ramsolutions.sw.magik.checks.MagikCheck;
 import org.sonar.check.Rule;
@@ -35,17 +36,15 @@ public class WarnedCallCheck extends MagikCheck {
 
   @Override
   protected void walkPreMethodInvocation(final AstNode node) {
-    final AstNode identifierNode = node.getFirstChild(MagikGrammar.IDENTIFIER);
-    if (identifierNode == null) {
+    final MethodInvocationNodeHelper helper = new MethodInvocationNodeHelper(node);
+    final String methodName = helper.getMethodName();
+    if (!this.getWarnedCalls().contains("." + methodName)) {
       return;
     }
 
-    final String identifier = "." + identifierNode.getTokenValue();
-    if (!this.getWarnedCalls().contains(identifier)) {
-      return;
-    }
-
-    this.addIssue(node, MESSAGE);
+    final AstNode methodNameNode = helper.getMethodNameNode();
+    final String message = String.format(MESSAGE, methodName);
+    this.addIssue(methodNameNode, message);
   }
 
   @Override