+++ magik-checks/src/test/java/nl/ramsolutions/sw/magik/checks/checks/ForbiddenCallCheckTest.java
@@ -5,48 +5,29 @@ import static org.assertj.core.api.Assertions.assertThat;
 import java.util.List;
 import nl.ramsolutions.sw.magik.checks.MagikCheck;
 import nl.ramsolutions.sw.magik.checks.MagikIssue;
-import org.junit.jupiter.api.Test;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.ValueSource;
 
 /** Test ForbiddenCallCheck. */
 class ForbiddenCallCheckTest extends MagikCheckTestBase {
 
-  @Test
-  void testProcedureDoSomething() {
+  @ParameterizedTest
+  @ValueSource(
+      strings = {
+        "do_something(1)",
+        "sw:do_something(1)",
+      })
+  void testValid(final String code) {
     final MagikCheck check = new ForbiddenCallCheck();
-    final String code = "do_something(1)";
     final List<MagikIssue> issues = this.runCheck(code, check);
     assertThat(issues).isEmpty();
   }
 
-  @Test
-  void testProcedureSwDoSomething() {
+  @ParameterizedTest
+  @ValueSource(
+      strings = {"show(1)", "sw:show(1)", "a.sys!slot(:a)", "a.sys!perform(:|xyz()|, x, y, z)"})
+  void testInvalid(final String code) {
     final MagikCheck check = new ForbiddenCallCheck();
-    final String code = "sw:do_something(1)";
-    final List<MagikIssue> issues = this.runCheck(code, check);
-    assertThat(issues).isEmpty();
-  }
-
-  @Test
-  void testProcedureShow() {
-    final MagikCheck check = new ForbiddenCallCheck();
-    final String code = "show(1)";
-    final List<MagikIssue> issues = this.runCheck(code, check);
-    assertThat(issues).hasSize(1);
-  }
-
-  @Test
-  void testProcedureSwShow() {
-    final MagikCheck check = new ForbiddenCallCheck();
-    final String code = "sw:show(1)";
-    final List<MagikIssue> issues = this.runCheck(code, check);
-    assertThat(issues).hasSize(1);
-  }
-
-  @Test
-  void testMethodP() {
-    final ForbiddenCallCheck check = new ForbiddenCallCheck();
-    check.forbiddenCalls = ".p";
-    final String code = "1.p";
     final List<MagikIssue> issues = this.runCheck(code, check);
     assertThat(issues).hasSize(1);
   }