+++ CHANGES.md
@@ -20,6 +20,7 @@
 - Fix UseValueCompare check to handle other notations for numbers + handle floats.
 - Fix magik-lint/magik-typed-lint not using located configuration file when showing checks.
 - Make check properties more descriptive by adding `max-` if needed.
+- Fix WarnedCallCheck and ForbiddenCallCheck to handle method invocations.
 - Several fixes.
 
 0.10.1 (2024-08-14)